chore: remove export of AnimatedHeaderHeightContext, move context to useAnimatedHeaderHeight.tsx
