+++ packages/native-stack/src/index.tsx
@@ -11,7 +11,6 @@ export { NativeStackView } from './views/NativeStackView';
 /**
  * Hooks
  */
-export { AnimatedHeaderHeightContext } from './utils/AnimatedHeaderHeightContext';
 export { useAnimatedHeaderHeight } from './utils/useAnimatedHeaderHeight';
 
 /**
