+++ packages/native-stack/src/views/NativeStackView.native.tsx
@@ -36,7 +36,7 @@ import type {
   NativeStackNavigationHelpers,
   NativeStackNavigationOptions,
 } from '../types';
-import { AnimatedHeaderHeightContext } from '../utils/AnimatedHeaderHeightContext';
+import { AnimatedHeaderHeightContext } from '../utils/useAnimatedHeaderHeight';
 import { useDismissedRouteError } from '../utils/useDismissedRouteError';
 import { useInvalidPreventRemoveError } from '../utils/useInvalidPreventRemoveError';
 import { DebugContainer } from './DebugContainer';
